fix: stale chile script process
